requirements.txt
streamlit==1.38.0
pandas==2.2.2
pydantic==2.8.2
langchain==0.2.14
langchain-openai==0.1.22
langchain-experimental==0.0.64
scikit-learn==1.5.1
xgboost==2.1.1
tensorflow==2.17.0
numpy==1.26.4
python-dotenv==1.0.1
plotly==5.22.0
pdfkit==1.0.0